[level 2] Title: n^2 배열 자르기, Time: 7.95 ms, Memory: 95.2 MB -BaekjoonHub

+++ 프로그래머스/2/87390. n＾2 배열 자르기/n＾2 배열 자르기.java
@@ -2,8 +2,7 @@ import java.util.*;
 
 class Solution {
     public int[] solution(int n, long left, long right) {
-        int[] answer = new int[(int)(right - left + 1)];
-
+        int[] answer = new int[(int)right - (int)left + 1];
         int count = 0;
         for (long i = left; i<=right; i++) {
             long x = i / n;
@@ -11,7 +10,6 @@ class Solution {
 
             answer[count++] = (int)Math.max(x, y) + 1;
         }
-        
         return answer;
     }
 }
